refactor: move clipboard into dir
