+++ src/modules/table-clipboard/index.ts
@@ -1,0 +1,1 @@
+export * from './paste-cell-into-cell';
